refactor: :recycle: разорвать цикл мобильных разрешений
Вынести перечисления, defaults и преобразование скалярных значений в ациклические модули.

Регистрировать обработчики через существующую точку регистрации мобильных свойств.
