+++ packages/core/metadata/appliedObjects/configuration/mobileApplicationPermissionDescription.ts
@@ -1,33 +1,72 @@
-import { importI8nTextFromXML } from "../../commonObjects/i8nText/fromXML"
-import { importI8nTextFromYAML } from "../../commonObjects/i8nText/fromYAML"
-import { exportI8nTextToXML } from "../../commonObjects/i8nText/toXML"
-import { exportI8nTextToYAML } from "../../commonObjects/i8nText/toYAML"
-import {
-  i8nTextRule,
-  type I8nText,
-  type I8nTextXML,
-  type I8nTextYAML,
-} from "../../commonObjects/i8nText/types"
-import type { ConfigurationContext } from "../../context/types"
+export interface MobileApplicationPermissionContext {
+  defaultLanguage: string
+}
 
-const descriptionRule = i8nTextRule({ preserveEmptyXML: true })
+export interface MobileApplicationPermissionDescription {
+  items: Record<string, string>
+}
+
+export type MobileApplicationPermissionDescriptionYAML = string | Record<string, string>
+
+export const MobileApplicationPermissionDescriptionJSONSchema = Type.Union([
+  Type.String(),
+  Type.Record(Type.String(), Type.String()),
+])
+
+export interface MobileApplicationPermissionDescriptionXML {
+  "v8:item"?:
+    | { "v8:lang": string; "v8:content": string }
+    | Array<{ "v8:lang": string; "v8:content": string }>
+}
+
+const normalizeArray = <T>(value: T | T[] | undefined): T[] => {
+  if (value === undefined) return []
+  return Array.isArray(value) ? value : [value]
+}
 
 export const importMobileApplicationPermissionDescriptionFromXML = (
-  context: ConfigurationContext,
-  value: I8nTextXML | ""
-): I8nText => importI8nTextFromXML(context, descriptionRule, value) ?? { items: {} }
+  _context: MobileApplicationPermissionContext,
+  value: MobileApplicationPermissionDescriptionXML | ""
+): MobileApplicationPermissionDescription => ({
+  items:
+    value === ""
+      ? {}
+      : Object.fromEntries(
+          normalizeArray(value["v8:item"]).map((item) => [item["v8:lang"], String(item["v8:content"] ?? "")])
+        ),
+})
 
 export const exportMobileApplicationPermissionDescriptionToXML = (
-  context: ConfigurationContext,
-  value: I8nText
-): I8nTextXML => exportI8nTextToXML(context, descriptionRule, value) ?? {}
+  context: MobileApplicationPermissionContext,
+  value: MobileApplicationPermissionDescription
+): MobileApplicationPermissionDescriptionXML => {
+  const entries = Object.entries(value.items)
+  const isEmpty = entries.every(([language, content]) => language === context.defaultLanguage && content === "")
+  if (isEmpty) return {}
+
+  const items = entries.map(([language, content]) => ({
+    "v8:lang": language,
+    "v8:content": content,
+  }))
+  return { "v8:item": items }
+}
 
 export const importMobileApplicationPermissionDescriptionFromYAML = (
-  context: ConfigurationContext,
-  value: I8nTextYAML
-): I8nText => importI8nTextFromYAML({ context, rule: descriptionRule, value }) ?? { items: {} }
+  context: MobileApplicationPermissionContext,
+  value: MobileApplicationPermissionDescriptionYAML
+): MobileApplicationPermissionDescription => ({
+  items: typeof value === "string" ? { [context.defaultLanguage]: value } : value,
+})
 
 export const exportMobileApplicationPermissionDescriptionToYAML = (
-  context: ConfigurationContext,
-  value: I8nText
-): I8nTextYAML => exportI8nTextToYAML({ context, rule: descriptionRule, value }) ?? ""
+  context: MobileApplicationPermissionContext,
+  value: MobileApplicationPermissionDescription
+): MobileApplicationPermissionDescriptionYAML => {
+  const languages = Object.keys(value.items)
+  if (languages.length === 0) return ""
+  if (languages.length === 1 && value.items[context.defaultLanguage] !== undefined) {
+    return value.items[context.defaultLanguage]
+  }
+  return value.items
+}
+import { Type } from "typebox"
