+++ packages/core/metadata/systemEnumerations/types.ts
@@ -2,6 +2,11 @@
 
 import { definePropertyRule, type ExactRuleParams } from "../commonObjects/ruleBuilder"
 import type { BasePropertyRule } from "../orchestration/property/types"
+import type {
+  MobileApplicationFunctionalities,
+  RequiredMobileApplicationPermissionMessages,
+  RequiredMobileApplicationPermissions,
+} from "../appliedObjects/configuration/mobileApplicationPermissionsEnumerations"
 
 export const DynamicListSearchStringViewModeToYAML = {
   Auto: "Авто",
@@ -2552,230 +2557,7 @@ export const MessageStatusFromYAML = {
 export type MessageStatus = keyof typeof MessageStatusToYAML
 export type MessageStatusYAML = keyof typeof MessageStatusFromYAML
 
-export const RequiredMobileApplicationPermissionMessagesToYAML = {
-  Biometrics: "Биометрия",
-  Location: "Геопозиционирование",
-  BackgroundLocation: "ГеопозиционированиеВФоновомРежиме",
-  BluetoothPrinters: "BluetoothПринтеры",
-  Contacts: "Контакты",
-  Calendars: "Календари",
-  NumberDialing: "НаборНомера",
-  CallProcessing: "ОбработкаЗвонков",
-  CallLog: "ЖурналЗвонков",
-  AutoSendSMS: "АвтоматическаяОтправкаSMSСообщений",
-  ReceiveSMS: "ПолучениеSMS",
-  SMSLog: "ЖурналSMS",
-  Camera: "Камера",
-  Microphone: "Микрофон",
-  MusicLibrary: "БиблиотекаМузыки",
-  PictureAndVideoLibraries: "БиблиотекиКартинокИВидео",
-  AudioPlaybackAndVibration: "ВоспроизведениеАудиоИВибрация",
-  PermissionGroupPhone: "PermissionGroupPhone",
-  PermissionGroupCallLog: "PermissionGroupCallLog",
-  PermissionGroupSMS: "PermissionGroupSMS",
-  PostNotifications: "PostNotifications",
-} as const
-
-export const RequiredMobileApplicationPermissionMessagesFromYAML = {
-  Биометрия: "Biometrics",
-  Геопозиционирование: "Location",
-  ГеопозиционированиеВФоновомРежиме: "BackgroundLocation",
-  BluetoothПринтеры: "BluetoothPrinters",
-  Контакты: "Contacts",
-  Календари: "Calendars",
-  НаборНомера: "NumberDialing",
-  ОбработкаЗвонков: "CallProcessing",
-  ЖурналЗвонков: "CallLog",
-  АвтоматическаяОтправкаSMSСообщений: "AutoSendSMS",
-  ПолучениеSMS: "ReceiveSMS",
-  ЖурналSMS: "SMSLog",
-  Камера: "Camera",
-  Микрофон: "Microphone",
-  БиблиотекаМузыки: "MusicLibrary",
-  БиблиотекиКартинокИВидео: "PictureAndVideoLibraries",
-  ВоспроизведениеАудиоИВибрация: "AudioPlaybackAndVibration",
-  PermissionGroupPhone: "PermissionGroupPhone",
-  PermissionGroupCallLog: "PermissionGroupCallLog",
-  PermissionGroupSMS: "PermissionGroupSMS",
-  PostNotifications: "PostNotifications",
-} as const
-
-export type RequiredMobileApplicationPermissionMessages =
-  keyof typeof RequiredMobileApplicationPermissionMessagesToYAML
-export type RequiredMobileApplicationPermissionMessagesYAML =
-  keyof typeof RequiredMobileApplicationPermissionMessagesFromYAML
-
-export const RequiredMobileApplicationPermissionsToYAML = {
-  Multimedia: "Мультимедиа",
-  Location: "Геопозиционирование",
-  Contacts: "Контакты",
-  Calendars: "Календари",
-  Telephony: "Телефония",
-  PushNotification: "PushУведомления",
-  LocalNotification: "ЛокальныеУведомления",
-  Print: "Печать",
-  InAppPurchases: "ВстроенныеПокупки",
-  Ads: "Реклама",
-  BackgroundLocation: "ГеопозиционированиеВФоновомРежиме",
-  BackgroundAudioPlayback: "ВоспроизведениеАудиоИВибрацияВФоновомРежиме",
-  FileExchangeWithPersonalComputer: "ОбменФайламиСПерсональнымКомпьютером",
-  CallPhone: "НаборНомера",
-  HandlePhoneCalls: "ОбработкаЗвонков",
-  CallLog: "ЖурналЗвонков",
-  SendSMS: "АвтоматическаяОтправкаSMSСообщений",
-  ReceiveSMS: "ПолучениеSMS",
-  SMSLog: "ЖурналSMS",
-  Camera: "Камера",
-  Microphone: "Микрофон",
-  MusicLibrary: "БиблиотекаМузыки",
-  PicturesAndVideoLibraries: "БиблиотекиКартинокИВидео",
-  AudioAndVibrationPlayback: "ВоспроизведениеАудиоИВибрация",
-  PermissionGroupPhone: "PermissionGroupPhone",
-  PermissionGroupCallLog: "PermissionGroupCallLog",
-  PermissionGroupSMS: "PermissionGroupSMS",
-  InstallPackages: "УстановкаПриложений",
-  AllowOSBackup: "РезервноеКопированиеСредствамиОС",
-  Biometrics: "Биометрия",
-  BluetoothPrinters: "BluetoothПринтеры",
-  WiFiPrinters: "WiFiПринтеры",
-  AllFilesAccess: "ДоступКоВсемФайлам",
-  Videoconferences: "Видеоконференции",
-  NFC: "NFC",
-  PostNotifications: "PostNotifications",
-  Geofences: "Геозоны",
-  IncomingShareRequests: "ВходящиеЗапросыПоделиться",
-  AllIncomingShareRequestsTypesProcessing: "ОбработкаВсехТиповВходящихЗапросовПоделиться",
-} as const
-
-export const RequiredMobileApplicationPermissionsFromYAML = {
-  Мультимедиа: "Multimedia",
-  Геопозиционирование: "Location",
-  Контакты: "Contacts",
-  Календари: "Calendars",
-  Телефония: "Telephony",
-  PushУведомления: "PushNotification",
-  ЛокальныеУведомления: "LocalNotification",
-  Печать: "Print",
-  ВстроенныеПокупки: "InAppPurchases",
-  Реклама: "Ads",
-  ГеопозиционированиеВФоновомРежиме: "BackgroundLocation",
-  ВоспроизведениеАудиоИВибрацияВФоновомРежиме: "BackgroundAudioPlayback",
-  ОбменФайламиСПерсональнымКомпьютером: "FileExchangeWithPersonalComputer",
-  НаборНомера: "CallPhone",
-  ОбработкаЗвонков: "HandlePhoneCalls",
-  ЖурналЗвонков: "CallLog",
-  АвтоматическаяОтправкаSMSСообщений: "SendSMS",
-  ПолучениеSMS: "ReceiveSMS",
-  ЖурналSMS: "SMSLog",
-  Камера: "Camera",
-  Микрофон: "Microphone",
-  БиблиотекаМузыки: "MusicLibrary",
-  БиблиотекиКартинокИВидео: "PicturesAndVideoLibraries",
-  ВоспроизведениеАудиоИВибрация: "AudioAndVibrationPlayback",
-  PermissionGroupPhone: "PermissionGroupPhone",
-  PermissionGroupCallLog: "PermissionGroupCallLog",
-  PermissionGroupSMS: "PermissionGroupSMS",
-  УстановкаПриложений: "InstallPackages",
-  РезервноеКопированиеСредствамиОС: "AllowOSBackup",
-  Биометрия: "Biometrics",
-  BluetoothПринтеры: "BluetoothPrinters",
-  WiFiПринтеры: "WiFiPrinters",
-  ДоступКоВсемФайлам: "AllFilesAccess",
-  Видеоконференции: "Videoconferences",
-  NFC: "NFC",
-  PostNotifications: "PostNotifications",
-  Геозоны: "Geofences",
-  ВходящиеЗапросыПоделиться: "IncomingShareRequests",
-  ОбработкаВсехТиповВходящихЗапросовПоделиться: "AllIncomingShareRequestsTypesProcessing",
-} as const
-
-export type RequiredMobileApplicationPermissions = keyof typeof RequiredMobileApplicationPermissionsToYAML
-export type RequiredMobileApplicationPermissionsYAML = keyof typeof RequiredMobileApplicationPermissionsFromYAML
-
-export const MobileApplicationFunctionalitiesToYAML = {
-  BluetoothPrinters: "BluetoothПринтеры",
-  NFC: "NFC",
-  PushNotifications: "PushУведомления",
-  WiFiPrinters: "WiFiПринтеры",
-  AutoSendSMS: "АвтоматическаяОтправкаSMSСообщений",
-  MusicLibrary: "БиблиотекаМузыки",
-  PictureAndVideoLibraries: "БиблиотекиКартинокИВидео",
-  Biometrics: "Биометрия",
-  Videoconferences: "Видеоконференции",
-  AudioPlaybackAndVibration: "ВоспроизведениеАудиоИВибрация",
-  BackgroundAudioPlaybackAndVibration: "ВоспроизведениеАудиоИВибрацияВФоновомРежиме",
-  InAppPurchases: "ВстроенныеПокупки",
-  IncomingShareRequests: "ВходящиеЗапросыПоделиться",
-  Geofences: "Геозоны",
-  Location: "Геопозиционирование",
-  BackgroundLocation: "ГеопозиционированиеВФоновомРежиме",
-  AllFilesAccess: "ДоступКоВсемФайлам",
-  SMSLog: "ЖурналSMS",
-  CallLog: "ЖурналЗвонков",
-  BackgroundAudioRecording: "ЗаписьАудиоВФоновомРежиме",
-  Calendars: "Календари",
-  Camera: "Камера",
-  Contacts: "Контакты",
-  LocalNotifications: "ЛокальныеУведомления",
-  Microphone: "Микрофон",
-  NumberDialing: "НаборНомера",
-  PersonalComputerFileExchange: "ОбменФайламиСПерсональнымКомпьютером",
-  AllIncomingShareRequestsTypesProcessing: "ОбработкаВсехТиповВходящихЗапросовПоделиться",
-  CallProcessing: "ОбработкаЗвонков",
-  ReceiveSMS: "ПолучениеSMS",
-  SpeechToText: "РаспознаваниеРечи",
-  OSBackup: "РезервноеКопированиеСредствамиОС",
-  Ads: "Реклама",
-  TextToSpeech: "СинтезРечи",
-  DocumentScanning: "СканированиеДокументов",
-  BarcodeScanning: "СканированиеШтрихКодов",
-  ApplicationUsageStatistics: "СтатистикаИспользованияПриложения",
-  InstallPackages: "УстановкаПриложений",
-} as const
-
-export const MobileApplicationFunctionalitiesFromYAML = {
-  BluetoothПринтеры: "BluetoothPrinters",
-  NFC: "NFC",
-  PushУведомления: "PushNotifications",
-  WiFiПринтеры: "WiFiPrinters",
-  АвтоматическаяОтправкаSMSСообщений: "AutoSendSMS",
-  БиблиотекаМузыки: "MusicLibrary",
-  БиблиотекиКартинокИВидео: "PictureAndVideoLibraries",
-  Биометрия: "Biometrics",
-  Видеоконференции: "Videoconferences",
-  ВоспроизведениеАудиоИВибрация: "AudioPlaybackAndVibration",
-  ВоспроизведениеАудиоИВибрацияВФоновомРежиме: "BackgroundAudioPlaybackAndVibration",
-  ВстроенныеПокупки: "InAppPurchases",
-  ВходящиеЗапросыПоделиться: "IncomingShareRequests",
-  Геозоны: "Geofences",
-  Геопозиционирование: "Location",
-  ГеопозиционированиеВФоновомРежиме: "BackgroundLocation",
-  ДоступКоВсемФайлам: "AllFilesAccess",
-  ЖурналSMS: "SMSLog",
-  ЖурналЗвонков: "CallLog",
-  ЗаписьАудиоВФоновомРежиме: "BackgroundAudioRecording",
-  Календари: "Calendars",
-  Камера: "Camera",
-  Контакты: "Contacts",
-  ЛокальныеУведомления: "LocalNotifications",
-  Микрофон: "Microphone",
-  НаборНомера: "NumberDialing",
-  ОбменФайламиСПерсональнымКомпьютером: "PersonalComputerFileExchange",
-  ОбработкаВсехТиповВходящихЗапросовПоделиться: "AllIncomingShareRequestsTypesProcessing",
-  ОбработкаЗвонков: "CallProcessing",
-  ПолучениеSMS: "ReceiveSMS",
-  РаспознаваниеРечи: "SpeechToText",
-  РезервноеКопированиеСредствамиОС: "OSBackup",
-  Реклама: "Ads",
-  СинтезРечи: "TextToSpeech",
-  СканированиеДокументов: "DocumentScanning",
-  СканированиеШтрихКодов: "BarcodeScanning",
-  СтатистикаИспользованияПриложения: "ApplicationUsageStatistics",
-  УстановкаПриложений: "InstallPackages",
-} as const
-
-export type MobileApplicationFunctionalities = keyof typeof MobileApplicationFunctionalitiesToYAML
-export type MobileApplicationFunctionalitiesYAML = keyof typeof MobileApplicationFunctionalitiesFromYAML
+export * from "../appliedObjects/configuration/mobileApplicationPermissionsEnumerations"
 
 export const NumericValueTypeToYAML = {
   Cardinal: "Количественное",
